Delete the correct website when deleting a tenant
When calling first() on the website object the first website in the database was being retrieved (it is not the relationship that was being accessed, so all websites could be accessed). This results in the wrong website being deleted when more than one tenant had been created and the more recent tenant deleted.